+++ trunk/OgreMain/src/OgreMatrix4.cpp
@@ -50,7 +50,8 @@ namespace Ogre
           0,    0,  0,   1);
 
     inline Real
-        MINOR(const Matrix4& m, const int r0, const int r1, const int r2, const int c0, const int c1, const int c2)
+        MINOR(const Matrix4& m, const size_t r0, const size_t r1, const size_t r2, 
+								const size_t c0, const size_t c1, const size_t c2)
     {
         return m[r0][c0] * (m[r1][c1] * m[r2][c2] - m[r2][c1] * m[r1][c2]) -
             m[r0][c1] * (m[r1][c0] * m[r2][c2] - m[r2][c0] * m[r1][c2]) +
